A bit of cleanup for loader + make CpgOverlayApplier private
