+++ semanticcpg/src/main/scala/io/shiftleft/layers/EnhancementRunner.scala
@@ -5,7 +5,7 @@ import io.shiftleft.codepropertygraph.Cpg
 
 class EnhancementRunner {
   def run(cpg: Cpg, serializedCpg: SerializedCpg): Unit = {
-    val language = cpg.metaData.language.head
+    val language = cpg.metaData.language.headOption.getOrElse("")
     new EnhancedBaseCreator(cpg.graph, language, serializedCpg).create
   }
 }
